+++ click_extended/decorators/check/__init__.py
@@ -3,11 +3,13 @@
 from click_extended.decorators.check.ends_with import ends_with
 from click_extended.decorators.check.exclusive import exclusive
 from click_extended.decorators.check.length import length
+from click_extended.decorators.check.not_empty import not_empty
 from click_extended.decorators.check.starts_with import starts_with
 
 __all__ = [
     "ends_with",
     "exclusive",
     "length",
+    "not_empty",
     "starts_with",
 ]