+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 - **Code Snippets**: Python code snippets.
 - **Dependencies**: Added `PyYAML` and `tomli` (only before 3.11) as a project dependency.
 - **Dev Dependencies**: Added `types-PyYAML` as a project dev dependency.
-- **`@starts_with`**: Child decorator to check if a strings starts with one or more patterns.
-- **`@ends_with`**: Child decorator to check if a strings ends with one or more patterns.
+- **`@not_empty`**: Child decorator to check if a string is empty.
+- **`@starts_with`**: Child decorator to check if a string starts with one or more patterns.
+- **`@ends_with`**: Child decorator to check if a string ends with one or more patterns.
 - **`@length`**: Child decorator to check if a string is within length bounds.
 - **`@less_than`**: Child decorator to check is a value is less or equal (inclusive=True) to a threshold.
 - **`@greater_than`**: Child decorator to check is a value is greater or equal (inclusive=True) to a threshold.